etcd api spec; rename lib/etcd/client_patch.rb to not break requires with rspec
