+++ agent/lib/etcd/client_patch.rb
